REFACTOR Moving all get neighbors into it's own folder
