+++ PuzzleSolverProject/GetNeighbors/UpLeftGetNeighborsFrom.cs
@@ -5,9 +5,9 @@ using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PuzzleSolverProject
+namespace PuzzleSolverProject.GetNeighbors
 {
-    class UpRightGetNeighborsFrom : IGetNeighborsFrom
+    class UpLeftGetNeighborsFrom : IGetNeighborsFrom
     {
         private const int STARTING_OFFSET = 0;
 
@@ -20,16 +20,16 @@ namespace PuzzleSolverProject
 
         public List<Vector2> GetNeighborsFrom(Vector2 startPosition, int length)
         {
-            List<Vector2> positionsUpRightFromStartPosition = new List<Vector2>();
-            for (int x = STARTING_OFFSET, y = STARTING_OFFSET; x < length && y > -length; x++, y--)
+            List<Vector2> positionsUpLeftFromStartPosition = new List<Vector2>();
+            for (int x = STARTING_OFFSET, y = STARTING_OFFSET; x > -length && y > -length; x--, y--)
             {
-                Vector2 upRightNeighbor = new Vector2(startPosition.X + x, startPosition.Y + y);
-                if (letters.ContainsKey(upRightNeighbor))
+                Vector2 UpLeftNeighbor = new Vector2(startPosition.X + x, startPosition.Y + y);
+                if (letters.ContainsKey(UpLeftNeighbor))
                 {
-                    positionsUpRightFromStartPosition.Add(upRightNeighbor);
+                    positionsUpLeftFromStartPosition.Add(UpLeftNeighbor);
                 }
             }
-            return positionsUpRightFromStartPosition;
+            return positionsUpLeftFromStartPosition;
         }
     }
 }
