+++ PuzzleSolverProject/PuzzleSolverProject.csproj
@@ -44,18 +44,18 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="DirectionEnum.cs" />
-    <Compile Include="DownLeftGetNeighborsFrom.cs" />
-    <Compile Include="DownRightGetNeighborsFrom.cs" />
-    <Compile Include="IGetNeighborsFrom.cs" />
-    <Compile Include="DownGetNeighborsFrom.cs" />
-    <Compile Include="LeftGetNeighborsFrom.cs" />
+    <Compile Include="GetNeighbors\DownLeftGetNeighborsFrom.cs" />
+    <Compile Include="GetNeighbors\DownRightGetNeighborsFrom.cs" />
+    <Compile Include="GetNeighbors\IGetNeighborsFrom.cs" />
+    <Compile Include="GetNeighbors\DownGetNeighborsFrom.cs" />
+    <Compile Include="GetNeighbors\LeftGetNeighborsFrom.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="PuzzleFileParser.cs" />
-    <Compile Include="RightGetNeighborsFrom.cs" />
-    <Compile Include="UpGetNeighborsFrom.cs" />
-    <Compile Include="UpLeftGetNeighborsFrom.cs" />
-    <Compile Include="UpRightGetNeighborsFrom.cs" />
+    <Compile Include="GetNeighbors\RightGetNeighborsFrom.cs" />
+    <Compile Include="GetNeighbors\UpGetNeighborsFrom.cs" />
+    <Compile Include="GetNeighbors\UpLeftGetNeighborsFrom.cs" />
+    <Compile Include="GetNeighbors\UpRightGetNeighborsFrom.cs" />
     <Compile Include="WordSeachAlgorithm.cs" />
     <Compile Include="WordSearchPuzzle.cs" />
   </ItemGroup>
