+++ PuzzleSolverProject/WordSeachAlgorithm.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using PuzzleSolverProject.GetNeighbors;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
